SELinux: protect the "initialized" state
Several well known attacks rely on the fact that SELinux uses the
internal "initialized" state as a preliminary condition on denying
permissions.

This patch reworks the state into a form which can be protected by
__ro_after_init plus pmalloc.

The outcome is that the state is permanently set, once a policy is
loaded.

This would prevent reloading of policy, however it's quite unlikely
that the SELinux policy is reloaded, at least in certain applications,
such as Android phones.

This patch requires some more work, to better integrate the protection,
but it shows how to counter a very typical attack on SELinux, with
the use of pmalloc.

Signed-off-by: Igor Stoppa <igor.stoppa@huawei.com>