Protection of the SELinux "initialized" state - stable
This patch reworks the "initialized" state into a form which can be
protected by using both __ro_after_init and pmalloc, combined.

The outcome is that the state is permanently set, once a policy is
loaded.

This patch requires some more work, to better integrate the protection,
but it shows how to counter a very typical attack on SELinux, with
the use of pmalloc.

Signed-off-by: Igor Stoppa <igor.stoppa@huawei.com>