Rename __safe_cmp to __safe_eval
The macro __safe_cmp is not really related to the cmp operation, even if
it was crafted for supporting it.

Instead, it can be used in much broader context, to prevent side effects
due to re-evaluating the parameters of a macro.

For this purpose, it can be renamed to a more neutral "__safe_eval", which
is what it really does.

Signed-off-by: Igor Stoppa <igor.stoppa@huawei.com>