Write Rare support for statically allocated memory
Support for allocating static memory that is normally read-only and can
be altered only by using specific primitives.

Static vairables declared with the __wr_after_init modifier are
normally writable until the completion of the init phase.

Afterwards, they can only be altered using a function of the write rare
family.

To minimize both accidental and malicious abuse of these functions, they
are inlined and they operate exclusively within a data segment, where
all the related data is allocated.

Reading this sort of data doesn't cause any of penalty, however the
write operation is intrinsically slower, due to the write mechanism.

When data in a page must be altered, the mapping of the page is not
affected, but the page is remapped to a new temporary random address, with
write permission.

This is meant to avoid that an attacker, busy looping on another core or
thread, can exploit a window of opportunity where the original mapping
would become writeable.

The attacker might not be interested in altering the data beign modified
by the OS, but rather, some other data which happens to be located in
the same memory page.

Of course the attack can still happen on the temporary mapping, but the
attacker needs to figure out where it is.
Ideally, every other core would be disabled during the write operation,
without a need for the remapping, however this would most likely cause
unacceptable freeze in the system.

Alternatively, the attack could be performed by carefully timing an
interrupt on the core executing the write operation at the remapped
address.
The attacker would freeze the writable mapping and have more time to
locate it. For this reason, local interrupts are disabled before the
temporary mapping is created, until it is destroyed.
The downside of this approach is that in the worst case an interrupt will
be delayed for the duration of a remapping/write/unmap cycle.
To provide an upper bound to such delay, one page at a time is altered,
re-enabling interrupts inbetween cycles, when the temporary mappings are
not available.

Such approach is sub-optimal from the perspective of altering large
amounts of data, which could be remapped in larger portions, however the
functionality is meant for data that is altered seldomly and can
tolerate somewhat slow updates.

To ensure integrity when writing multi-byte data that might cause
problems if not modified atomically (ex: pointers), write_once is used,
however there might be problems if used with packed structures, where
the data alignment is not guarranteed. In these cases it might be
necessary to use some form of locking or RCU between readers and writers,
even if there is only one writer.

Regarding concurrent write rare calls that affect non-overlapping data
ranges, they can happen simultaneously, each at its own remapped
address, as long as the HW cache coherency guarrantees, at physical
level, that the symultaneous writes will be handled correctly.

Signed-off-by: Igor Stoppa <igor.stoppa@huawei.com>
CC: Ashish Raj <araj@elisanet.fi>
CC: Carlos Chinea Perez <carlos.chinea.perez@huawei.com>
CC: Remi Denis Courmont <remi.denis.courmont@huawei.com>